refactor(economy): extraction snapshot and non-GP shared splits (#4550)
Extract TileExtractionDisplayContribution into its own module, split
non_gp_extraction_shared into faction walk, tile contribution compute,
and tile walk siblings, and iterate improvementByTile once in the GP
aggregator while preserving display semantics.
