+++ packages/colonizethis_economy/lib/src/economy/province_extraction_snapshot_builder.dart
@@ -9,91 +9,16 @@ import 'package:colonizethis_world/colonizethis_world.dart';
 
 import 'economy_resource_constants.dart';
 import 'game_lookup_helpers.dart';
+import 'tile_extraction_display_contribution.dart';
 import 'tile_extraction_pipeline.dart';
-import 'tile_extraction_yield.dart';
+
+export 'tile_extraction_display_contribution.dart';
 
 /// Mutable per-commodity accumulator used while building a province snapshot.
 class _CommodityAcc {
   int effective = 0;
   int full = 0;
   final Set<String> tileKeys = <String>{};
-}
-
-/// Dual production/effective contribution for one improved extractable tile.
-class TileExtractionDisplayContribution {
-  const TileExtractionDisplayContribution({
-    required this.provinceId,
-    required this.commodityId,
-    required this.effective,
-    required this.full,
-    required this.tileKey,
-  });
-
-  final String provinceId;
-  final String commodityId;
-  final int effective;
-  final int full;
-  final String tileKey;
-}
-
-/// Computes production (full) and effective yield for one improved tile,
-/// including disconnected tiles where effective is 0 and full > 0.
-///
-/// Shares resolve/clamp/production with [computeTileYieldContribution] via
-/// [resolveImprovedTileProductionPrelude] (Refs #4014).
-TileExtractionDisplayContribution? computeTileExtractionDisplayContribution({
-  required Game game,
-  required Map<String, TileMapResult> tileMapByRegion,
-  required String tileKey,
-  required Set<String> connectedTileKeys,
-  required Map<String, int> pathTransportCap,
-  required Set<String> connectedByRoadRule,
-  required Set<String> portTileKeys,
-  required String? capitalProvinceId,
-  required int Function(CommodityId commodityId) techCapForCommodity,
-  required bool Function(String tileKey, CommodityId commodityId)
-  isCommodityExtractable,
-  Map<String, Province>? provincesByFullId,
-}) {
-  final prelude = resolveImprovedTileProductionPrelude(
-    game: game,
-    tileMapByRegion: tileMapByRegion,
-    tileKey: tileKey,
-    provincesByFullId: provincesByFullId,
-    logContext: 'provinceExtractionSnapshot',
-    techCapForCommodity: techCapForCommodity,
-    isCommodityExtractable: isCommodityExtractable,
-  );
-  if (prelude == null) {
-    return null;
-  }
-
-  var effective = 0;
-  if (connectedTileKeys.contains(tileKey)) {
-    final province = prelude.province;
-    final townTileKey = province.townTileKey;
-    final townTileIsPort =
-        townTileKey != null && portTileKeys.contains(townTileKey);
-    effective = computeEffectiveTileYield(
-      tileState: game.worldState.tileState,
-      tileKey: tileKey,
-      techCap: prelude.techCap,
-      townDevelopmentCap: province.townDevelopmentLevel,
-      townTileIsPort: townTileIsPort,
-      isCapitalProvince: prelude.provinceId == capitalProvinceId,
-      usesRoadRule: connectedByRoadRule.contains(tileKey),
-      portTileKeys: portTileKeys,
-      pathTransportCap: pathTransportCap,
-    );
-  }
-
-  return TileExtractionDisplayContribution(
-    provinceId: prelude.provinceId,
-    commodityId: prelude.commodityId,
-    effective: effective,
-    full: prelude.production,
-    tileKey: tileKey,
-  );
 }
 
 /// Projects Extraction for a single province from the current post-resolution
@@ -131,11 +56,29 @@ Map<String, ProvinceExtractionSnapshot> computeProvinceExtractionSnapshots({
 }) {
   final provincesByFullId = buildProvinceIndex(game);
   final portTileKeys = collectPortTileKeys(game);
+  final playersById = {for (final player in game.players) player.id: player};
   final accByProvince = <String, Map<String, _CommodityAcc>>{};
   final ownerByProvince = <String, String>{};
   final capitalBonusByProvince = <String, int>{};
 
-  for (final player in game.players) {
+  for (final tileKey
+      in game.worldState.tileState.improvementByTile.keys) {
+    final tileContext = resolveTileKeyExtractionContext(
+      tileKey: tileKey,
+      tileMapByRegion: tileMapByRegion,
+      provincesByFullId: provincesByFullId,
+      game: game,
+      logContext: 'provinceExtractionSnapshot',
+    );
+    if (tileContext == null) {
+      continue;
+    }
+
+    final player = playersById[tileContext.province.ownerId];
+    if (player == null) {
+      continue;
+    }
+
     final cr = connectivityResult[player.id];
     final connected = cr?.connected ?? const <String>{};
     final pathTransportCap = cr?.pathTransportCap ?? const <String, int>{};
@@ -143,47 +86,42 @@ Map<String, ProvinceExtractionSnapshot> computeProvinceExtractionSnapshots({
     final prospected =
         game.worldState.playerProspectedTiles[player.id] ?? const <String>{};
 
-    for (final tileKey in game.worldState.tileState.improvementByTile.keys) {
-      final contribution = computeTileExtractionDisplayContribution(
-        game: game,
-        tileMapByRegion: tileMapByRegion,
-        tileKey: tileKey,
-        connectedTileKeys: connected,
-        pathTransportCap: pathTransportCap,
-        connectedByRoadRule: roadRuleTiles,
-        portTileKeys: portTileKeys,
-        capitalProvinceId: player.capitalProvinceId,
-        provincesByFullId: provincesByFullId,
-        techCapForCommodity: (commodityId) =>
-            techCapForPlayerAndResource?.call(player.id, commodityId) ??
-            techCapForPlayer(player.id),
-        isCommodityExtractable: (tk, commodityId) =>
-            !kMineralResourceIds.contains(commodityId) ||
-            prospected.contains(tk),
-      );
-      if (contribution == null) {
-        continue;
-      }
-
-      final province = provincesByFullId[contribution.provinceId];
-      if (province == null || province.ownerId != player.id) {
-        continue;
-      }
-
-      ownerByProvince[contribution.provinceId] = player.id;
-      final byCommodity = accByProvince.putIfAbsent(
-        contribution.provinceId,
-        () => {},
-      );
-      final acc = byCommodity.putIfAbsent(
-        contribution.commodityId,
-        _CommodityAcc.new,
-      );
-      acc.effective += contribution.effective;
-      acc.full += contribution.full;
-      acc.tileKeys.add(contribution.tileKey);
+    final contribution = computeTileExtractionDisplayContribution(
+      game: game,
+      tileMapByRegion: tileMapByRegion,
+      tileKey: tileKey,
+      connectedTileKeys: connected,
+      pathTransportCap: pathTransportCap,
+      connectedByRoadRule: roadRuleTiles,
+      portTileKeys: portTileKeys,
+      capitalProvinceId: player.capitalProvinceId,
+      provincesByFullId: provincesByFullId,
+      techCapForCommodity: (commodityId) =>
+          techCapForPlayerAndResource?.call(player.id, commodityId) ??
+          techCapForPlayer(player.id),
+      isCommodityExtractable: (tk, commodityId) =>
+          !kMineralResourceIds.contains(commodityId) ||
+          prospected.contains(tk),
+    );
+    if (contribution == null) {
+      continue;
     }
 
+    ownerByProvince[contribution.provinceId] = player.id;
+    final byCommodity = accByProvince.putIfAbsent(
+      contribution.provinceId,
+      () => {},
+    );
+    final acc = byCommodity.putIfAbsent(
+      contribution.commodityId,
+      _CommodityAcc.new,
+    );
+    acc.effective += contribution.effective;
+    acc.full += contribution.full;
+    acc.tileKeys.add(contribution.tileKey);
+  }
+
+  for (final player in game.players) {
     final capBonus = game.capitalTileGrainBonusPerTurn;
     final capitalProvinceId = player.capitalProvinceId;
     if (player.capitalTile != null &&
